JAVA-1869: Add ProgrammaticDriverConfigLoader
Motivation:

To enable a convenient way to configure a Driver Session
programmatically without being part of the top-level API.

Modifications:

Added ProgrammaticDriverConfigLoader which provides a builder for
passing in configuration programmatically.

Added class DriverOption back to they may be configured.

Updated SessionUtils to provide a way of constructing a Session
with the given ProgrammaticDriverConfigLoader.Builder so it can
be easily used in integration tests.  Updated SpeculativeExecutionIT
to use it.

Result:

ProgrammaticDriverConfigLoader.builder() may be used to construct
a DriverConfigLoader which may be passed in via
SessionBuilder.withConfigLoader.